requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.5.0
Django==4.1.3
django-bootstrap4==22.3
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
drf-spectacular==0.24.2
flake8==6.0.0
flower==1.2.0
gunicorn==20.1.0
humanize==4.4.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
mccabe==0.7.0
mysqlclient==2.1.1
packaging==21.3
Pillow==9.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.33
psutil==5.9.4
pycodestyle==2.10.0
pydocstyle==6.1.1
pyflakes==3.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
redis==4.3.5
requests==2.28.1
six==1.16.0
snowballstemmer==2.2.0
sorl-thumbnail==12.9.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stripe==5.0.0
tornado==6.2
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
